fix map layersccomponents rendering
